+++ ironflow/node_tools/__init__.py
@@ -25,7 +25,7 @@ Example:
     >>> gui.register_node(MyNode)
 """
 
-import ironflow.custom_nodes.input_widgets
+import ironflow.node_tools.input_widgets
 from ironflow.model import dtypes, NodeInputBP
 from ironflow.model.node import Node
 from ironflow.model.port import NodeOutputBP
